+++ models/sources/segment/base/base_segment_tracks.sql
@@ -1,0 +1,1 @@
+select * from {{ source('SEGMENT_LIVE', 'TRACKS') }}